Move Config to config.go
