+++ pkg/integrations/cloudwatch_exporter/config.go
@@ -1,8 +1,19 @@
 package cloudwatch_exporter
 
 import (
+	"crypto/md5"
+	"encoding/hex"
+	"fmt"
+	"time"
+
+	"github.com/go-kit/log"
 	yaceConf "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/config"
 	yaceModel "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/model"
+	"gopkg.in/yaml.v2"
+
+	"github.com/grafana/agent/pkg/integrations"
+	integrations_v2 "github.com/grafana/agent/pkg/integrations/v2"
+	"github.com/grafana/agent/pkg/integrations/v2/metricsutils"
 )
 
 const (
@@ -18,6 +29,106 @@ var addCloudwatchTimestamp = false
 
 // Avoid producing absence of values in metrics
 var nilToZero = true
+
+func init() {
+	integrations.RegisterIntegration(&Config{})
+	integrations_v2.RegisterLegacy(&Config{}, integrations_v2.TypeMultiplex, metricsutils.NewNamedShim("cloudwatch"))
+}
+
+// Config is the configuration for the CloudWatch metrics integration
+type Config struct {
+	STSRegion string          `yaml:"sts_region"`
+	Discovery DiscoveryConfig `yaml:"discovery"`
+	Static    []StaticJob     `yaml:"static"`
+}
+
+// DiscoveryConfig configures scraping jobs that will auto-discover metrics dimensions for a given service.
+type DiscoveryConfig struct {
+	ExportedTags TagsPerNamespace `yaml:"exported_tags"`
+	Jobs         []*DiscoveryJob  `yaml:"jobs"`
+}
+
+// TagsPerNamespace represents for each namespace, a list of tags that will be exported as labels in each metric.
+type TagsPerNamespace map[string][]string
+
+// DiscoveryJob configures a discovery job for a given service.
+type DiscoveryJob struct {
+	InlineRegionAndRoles `yaml:",inline"`
+	InlineCustomTags     `yaml:",inline"`
+	SearchTags           []Tag    `yaml:"search_tags"`
+	Type                 string   `yaml:"type"`
+	Metrics              []Metric `yaml:"metrics"`
+}
+
+// StaticJob will scrape metrics that match all defined dimensions.
+type StaticJob struct {
+	InlineRegionAndRoles `yaml:",inline"`
+	InlineCustomTags     `yaml:",inline"`
+	Name                 string      `yaml:"name"`
+	Namespace            string      `yaml:"namespace"`
+	Dimensions           []Dimension `yaml:"dimensions"`
+	Metrics              []Metric    `yaml:"metrics"`
+}
+
+// InlineRegionAndRoles exposes for each supported job, the AWS regions and IAM roles in which the agent should perform the
+// scrape.
+type InlineRegionAndRoles struct {
+	Regions []string `yaml:"regions"`
+	Roles   []Role   `yaml:"roles"`
+}
+
+type InlineCustomTags struct {
+	CustomTags []Tag `yaml:"custom_tags"`
+}
+
+type Role struct {
+	RoleArn    string `yaml:"role_arn"`
+	ExternalID string `yaml:"external_id"`
+}
+
+type Dimension struct {
+	Name  string `yaml:"name"`
+	Value string `yaml:"value"`
+}
+
+type Tag struct {
+	Key   string `yaml:"key"`
+	Value string `yaml:"value"`
+}
+
+type Metric struct {
+	Name       string        `yaml:"name"`
+	Statistics []string      `yaml:"statistics"`
+	Period     time.Duration `yaml:"period"`
+}
+
+// Name returns the name of the integration this config is for.
+func (c *Config) Name() string {
+	return "cloudwatch_exporter"
+}
+
+func (c *Config) InstanceKey(agentKey string) (string, error) {
+	return getHash(c)
+}
+
+// NewIntegration creates a new integration from the config.
+func (c *Config) NewIntegration(l log.Logger) (integrations.Integration, error) {
+	exporterConfig, err := ToYACEConfig(c)
+	if err != nil {
+		return nil, fmt.Errorf("invalid cloudwatch exporter configuration: %w", err)
+	}
+	return newCloudwatchExporter(c.Name(), l, exporterConfig), nil
+}
+
+// getHash calculates the MD5 hash of the yaml representation of the config
+func getHash(c *Config) (string, error) {
+	bytes, err := yaml.Marshal(c)
+	if err != nil {
+		return "", err
+	}
+	hash := md5.Sum(bytes)
+	return hex.EncodeToString(hash[:]), nil
+}
 
 // ToYACEConfig converts a Config into YACE's config model. Note that the conversion is not direct, some values
 // have been opinionated to simplify the config model the agent exposes for this integration.
