+++ pkg/integrations/cloudwatch_exporter/config_test.go
@@ -5,6 +5,7 @@ import (
 
 	yaceConf "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/config"
 	yaceModel "github.com/nerdswords/yet-another-cloudwatch-exporter/pkg/model"
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 )
@@ -145,4 +146,27 @@ func TestTranslateConfigToYACEConfig(t *testing.T) {
 	require.NoError(t, err, "failed to translate to YACE configuration")
 
 	require.EqualValues(t, expectedConfig, yaceConf)
+}
+
+func TestCloudwatchExporterConfigInstanceKey(t *testing.T) {
+	cfg1 := &Config{
+		STSRegion: "us-east-2",
+	}
+	cfg2 := &Config{
+		STSRegion: "us-east-3",
+	}
+
+	cfg1Hash, err := cfg1.InstanceKey("")
+	require.NoError(t, err)
+	cfg2Hash, err := cfg2.InstanceKey("")
+	require.NoError(t, err)
+
+	assert.NotEqual(t, cfg1Hash, cfg2Hash)
+
+	// test that making them equal in values leads to the same instance key
+	cfg2.STSRegion = "us-east-2"
+	cfg2Hash, err = cfg2.InstanceKey("")
+	require.NoError(t, err)
+
+	assert.Equal(t, cfg1Hash, cfg2Hash)
 }
